Support .per(value,unit) and similar as in doc
Specifically, the .per(), .plus(), .minus(), .times(), and .isSame()
methods all support either a single Quantity argument, or a pair of
(value,unit) arguments which are converted to a Quantity on the fly as
shown in the README example.

Also fixed syntax error in that README example, added example of a
predefined Quantity, added test cases for new features, and removed
the 'u' shortcut in the test cases - it isn't really useful now that
the various methods can create a Quantity on the fly. (This way the
test cases are written more like you would expect actual code to
be written.)